Create CheckSparseNo.java
Given a number N. The task is to check whether it is sparse or not. A number is said to be a sparse number if no two or more consecutive bits are set in the binary representation.
Note: Instead of right shift, we could have used left shift also, but left shift might lead to overflow in some cases.